refactor: split commands/ai tests into a sibling test file
The round-11 budget-outcome tests pushed commands/ai.rs past the R8
1400-LOC cap; same directory-module split as tools_quality, controller,
and ai_generations (mod.rs 1113 + test.rs 321, 19 tests before and
after, pure movement).
